Improved disabling the upgrades module
