+++ src/main/java/com/bgsoftware/superiorskyblock/commands/admin/AdminCommandsMap.java
@@ -14,12 +14,10 @@ public class AdminCommandsMap extends CommandsMap {
         clearCommands();
 
         registerCommand(new CmdAdminAdd());
-        registerCommand(new CmdAdminAddBlockLimit());
         registerCommand(new CmdAdminAddBonus());
         if (plugin.getSettings().isCoopMembers())
             registerCommand(new CmdAdminAddCoopLimit());
         registerCommand(new CmdAdminAddDisbands());
-        registerCommand(new CmdAdminAddEntityLimit());
         registerCommand(new CmdAdminAddSize());
         registerCommand(new CmdAdminAddTeamLimit());
         registerCommand(new CmdAdminAddWarpsLimit());
@@ -47,8 +45,6 @@ public class AdminCommandsMap extends CommandsMap {
         registerCommand(new CmdAdminPurge());
         registerCommand(new CmdAdminRecalc());
         registerCommand(new CmdAdminReload());
-        registerCommand(new CmdAdminRemoveBlockLimit());
-        registerCommand(new CmdAdminRemoveEntityLimit());
         registerCommand(new CmdAdminRemoveRatings());
         registerCommand(new CmdAdminResetPermissions());
         registerCommand(new CmdAdminResetSettings());
@@ -57,13 +53,11 @@ public class AdminCommandsMap extends CommandsMap {
         registerCommand(new CmdAdminSetBankLimit());
         registerCommand(new CmdAdminSetBiome());
         registerCommand(new CmdAdminSetBlockAmount());
-        registerCommand(new CmdAdminSetBlockLimit());
         registerCommand(new CmdAdminSetBonus());
         registerCommand(new CmdAdminSetChestRow());
         if (plugin.getSettings().isCoopMembers())
             registerCommand(new CmdAdminSetCoopLimit());
         registerCommand(new CmdAdminSetDisbands());
-        registerCommand(new CmdAdminSetEntityLimit());
         registerCommand(new CmdAdminSetIslandPreview());
         registerCommand(new CmdAdminSetLeader());
         registerCommand(new CmdAdminSetPermission());
