+++ src/main/java/com/bgsoftware/superiorskyblock/module/upgrades/commands/CmdAdminRemoveBlockLimit.java
@@ -1,4 +1,4 @@
-package com.bgsoftware.superiorskyblock.commands.admin;
+package com.bgsoftware.superiorskyblock.module.upgrades.commands;
 
 import com.bgsoftware.common.annotations.Nullable;
 import com.bgsoftware.superiorskyblock.SuperiorSkyblockPlugin;
@@ -7,54 +7,50 @@ import com.bgsoftware.superiorskyblock.api.key.Key;
 import com.bgsoftware.superiorskyblock.api.wrappers.SuperiorPlayer;
 import com.bgsoftware.superiorskyblock.commands.CommandTabCompletes;
 import com.bgsoftware.superiorskyblock.commands.IAdminIslandCommand;
-import com.bgsoftware.superiorskyblock.commands.arguments.CommandArguments;
-import com.bgsoftware.superiorskyblock.commands.arguments.NumberArgument;
-import com.bgsoftware.superiorskyblock.core.events.args.PluginEventArgs;
-import com.bgsoftware.superiorskyblock.core.events.plugin.PluginEvent;
 import com.bgsoftware.superiorskyblock.core.events.plugin.PluginEventsFactory;
 import com.bgsoftware.superiorskyblock.core.formatting.Formatters;
 import com.bgsoftware.superiorskyblock.core.key.Keys;
 import com.bgsoftware.superiorskyblock.core.messages.Message;
 import org.bukkit.command.CommandSender;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public class CmdAdminSetBlockLimit implements IAdminIslandCommand {
+public class CmdAdminRemoveBlockLimit implements IAdminIslandCommand {
 
     @Override
     public List<String> getAliases() {
-        return Collections.singletonList("setblocklimit");
+        return Arrays.asList("removeblocklimit", "remblocklimit");
     }
 
     @Override
     public String getPermission() {
-        return "superior.admin.setblocklimit";
+        return "superior.admin.removeblocklimit";
     }
 
     @Override
     public String getUsage(java.util.Locale locale) {
-        return "admin setblocklimit <" +
+        return "admin removeblocklimit <" +
                 Message.COMMAND_ARGUMENT_PLAYER_NAME.getMessage(locale) + "/" +
                 Message.COMMAND_ARGUMENT_ISLAND_NAME.getMessage(locale) + "/" +
                 Message.COMMAND_ARGUMENT_ALL_ISLANDS.getMessage(locale) + "> <" +
-                Message.COMMAND_ARGUMENT_MATERIAL.getMessage(locale) + "> <" +
-                Message.COMMAND_ARGUMENT_LIMIT.getMessage(locale) + ">";
+                Message.COMMAND_ARGUMENT_MATERIAL.getMessage(locale) + ">";
     }
 
     @Override
     public String getDescription(java.util.Locale locale) {
-        return Message.COMMAND_DESCRIPTION_ADMIN_SET_BLOCK_LIMIT.getMessage(locale);
+        return Message.COMMAND_DESCRIPTION_ADMIN_REMOVE_BLOCK_LIMIT.getMessage(locale);
     }
 
     @Override
     public int getMinArgs() {
-        return 5;
+        return 4;
     }
 
     @Override
     public int getMaxArgs() {
-        return 5;
+        return 4;
     }
 
     @Override
@@ -71,20 +67,12 @@ public class CmdAdminSetBlockLimit implements IAdminIslandCommand {
     public void execute(SuperiorSkyblockPlugin plugin, CommandSender sender, @Nullable SuperiorPlayer targetPlayer, List<Island> islands, String[] args) {
         Key key = Keys.ofMaterialAndData(args[3]);
 
-        NumberArgument<Integer> arguments = CommandArguments.getLimit(sender, args[4]);
-
-        if (!arguments.isSucceed())
-            return;
-
-        int limit = arguments.getNumber();
-
         int islandsChangedCount = 0;
 
         for (Island island : islands) {
-            PluginEvent<PluginEventArgs.IslandChangeBlockLimit> event = PluginEventsFactory.callIslandChangeBlockLimitEvent(island, sender, key, limit);
-            if (!event.isCancelled()) {
-                island.setBlockLimit(key, event.getArgs().blockLimit);
+            if (PluginEventsFactory.callIslandRemoveBlockLimitEvent(island, sender, key)) {
                 ++islandsChangedCount;
+                island.removeBlockLimit(key);
             }
         }
