+++ src/main/java/com/bgsoftware/superiorskyblock/module/upgrades/commands/CmdAdminSetEntityLimit.java
@@ -1,4 +1,4 @@
-package com.bgsoftware.superiorskyblock.commands.admin;
+package com.bgsoftware.superiorskyblock.module.upgrades.commands;
 
 import com.bgsoftware.common.annotations.Nullable;
 import com.bgsoftware.superiorskyblock.SuperiorSkyblockPlugin;
@@ -7,50 +7,54 @@ import com.bgsoftware.superiorskyblock.api.key.Key;
 import com.bgsoftware.superiorskyblock.api.wrappers.SuperiorPlayer;
 import com.bgsoftware.superiorskyblock.commands.CommandTabCompletes;
 import com.bgsoftware.superiorskyblock.commands.IAdminIslandCommand;
+import com.bgsoftware.superiorskyblock.commands.arguments.CommandArguments;
+import com.bgsoftware.superiorskyblock.commands.arguments.NumberArgument;
+import com.bgsoftware.superiorskyblock.core.events.args.PluginEventArgs;
+import com.bgsoftware.superiorskyblock.core.events.plugin.PluginEvent;
 import com.bgsoftware.superiorskyblock.core.events.plugin.PluginEventsFactory;
 import com.bgsoftware.superiorskyblock.core.formatting.Formatters;
 import com.bgsoftware.superiorskyblock.core.key.Keys;
 import com.bgsoftware.superiorskyblock.core.messages.Message;
 import org.bukkit.command.CommandSender;
 
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public class CmdAdminRemoveEntityLimit implements IAdminIslandCommand {
+public class CmdAdminSetEntityLimit implements IAdminIslandCommand {
 
     @Override
     public List<String> getAliases() {
-        return Arrays.asList("removeentitylimit", "rementitylimit");
+        return Collections.singletonList("setentitylimit");
     }
 
     @Override
     public String getPermission() {
-        return "superior.admin.removeentitylimit";
+        return "superior.admin.setentitylimit";
     }
 
     @Override
     public String getUsage(java.util.Locale locale) {
-        return "admin removeentitylimit <" +
+        return "admin setentitylimit <" +
                 Message.COMMAND_ARGUMENT_PLAYER_NAME.getMessage(locale) + "/" +
                 Message.COMMAND_ARGUMENT_ISLAND_NAME.getMessage(locale) + "/" +
                 Message.COMMAND_ARGUMENT_ALL_ISLANDS.getMessage(locale) + "> <" +
-                Message.COMMAND_ARGUMENT_ENTITY.getMessage(locale) + ">";
+                Message.COMMAND_ARGUMENT_ENTITY.getMessage(locale) + "> <" +
+                Message.COMMAND_ARGUMENT_LIMIT.getMessage(locale) + ">";
     }
 
     @Override
     public String getDescription(java.util.Locale locale) {
-        return Message.COMMAND_DESCRIPTION_ADMIN_REMOVE_ENTITY_LIMIT.getMessage(locale);
+        return Message.COMMAND_DESCRIPTION_ADMIN_SET_ENTITY_LIMIT.getMessage(locale);
     }
 
     @Override
     public int getMinArgs() {
-        return 4;
+        return 5;
     }
 
     @Override
     public int getMaxArgs() {
-        return 4;
+        return 5;
     }
 
     @Override
@@ -65,14 +69,23 @@ public class CmdAdminRemoveEntityLimit implements IAdminIslandCommand {
 
     @Override
     public void execute(SuperiorSkyblockPlugin plugin, CommandSender sender, @Nullable SuperiorPlayer targetPlayer, List<Island> islands, String[] args) {
-        Key key = Keys.ofEntityType(args[3]);
+        Key entityKey = Keys.ofEntityType(args[3]);
+
+        NumberArgument<Integer> arguments = CommandArguments.getLimit(sender, args[4]);
+
+        if (!arguments.isSucceed())
+            return;
+
+        int limit = arguments.getNumber();
 
         int islandsChangedCount = 0;
 
         for (Island island : islands) {
-            if (PluginEventsFactory.callIslandRemoveEntityLimitEvent(island, sender, key)) {
+            PluginEvent<PluginEventArgs.IslandChangeEntityLimit> event = PluginEventsFactory.callIslandChangeEntityLimitEvent(
+                    island, sender, entityKey, limit);
+            if (!event.isCancelled()) {
+                island.setEntityLimit(entityKey, event.getArgs().entityLimit);
                 ++islandsChangedCount;
-                island.removeEntityLimit(key);
             }
         }
 
@@ -80,11 +93,11 @@ public class CmdAdminRemoveEntityLimit implements IAdminIslandCommand {
             return;
 
         if (islandsChangedCount > 1)
-            Message.CHANGED_ENTITY_LIMIT_ALL.send(sender, Formatters.CAPITALIZED_FORMATTER.format(key.getGlobalKey()));
+            Message.CHANGED_ENTITY_LIMIT_ALL.send(sender, Formatters.CAPITALIZED_FORMATTER.format(entityKey.getGlobalKey()));
         else if (targetPlayer == null)
-            Message.CHANGED_ENTITY_LIMIT_NAME.send(sender, Formatters.CAPITALIZED_FORMATTER.format(key.getGlobalKey()), islands.get(0).getName());
+            Message.CHANGED_ENTITY_LIMIT_NAME.send(sender, Formatters.CAPITALIZED_FORMATTER.format(entityKey.getGlobalKey()), islands.get(0).getName());
         else
-            Message.CHANGED_ENTITY_LIMIT.send(sender, Formatters.CAPITALIZED_FORMATTER.format(key.getGlobalKey()), targetPlayer.getName());
+            Message.CHANGED_ENTITY_LIMIT.send(sender, Formatters.CAPITALIZED_FORMATTER.format(entityKey.getGlobalKey()), targetPlayer.getName());
     }
 
     @Override
